+++ Models/STT/tansformer_whisper.py
@@ -155,7 +155,7 @@ class TransformerWhisper(metaclass=SingletonMeta):
                         )
 
                 print(f"Loading Whisper-Transformer model: {model} on {device} with {compute_type} precision using {attention_type}...")
-                self.model = WhisperForConditionalGeneration.from_pretrained(str(Path(self.model_cache_path / model).resolve()), torch_dtype=compute_dtype, quantization_config=quantization_config, device_map=self.compute_device, attn_implementation=attention_type)
+                self.model = WhisperForConditionalGeneration.from_pretrained(str(Path(self.model_cache_path / model).resolve()), dtype=compute_dtype, quantization_config=quantization_config, device_map=self.compute_device, attn_implementation=attention_type)
                 #try:
                 #    # Enable static cache and compile the forward pass
                 #    self.model.generation_config.cache_implementation = "static"
